# internal/docs/swagger.yaml
basePath: /v1
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  core.RunDifferentialGamePoint:
    properties:
      cumulative_diff:
        description: running season sum
        type: integer
      date:
        description: YYYY-MM-DD
        type: string
      differential:
        description: single-game RS - RA
        type: integer
      game_id:
        type: string
      home:
        type: boolean
      opponent_id:
        type: string
      runs_allowed:
        type: integer
      runs_scored:
        type: integer
    type: object
  core.RunDifferentialSeries:
    properties:
      entity_id:
        description: TeamID
        type: string
      entity_type:
        description: usually "team"
        type: string
      games:
        description: Per-game series, ordered chronologically.
        items:
          $ref: '#/definitions/core.RunDifferentialGamePoint'
        type: array
      games_played:
        description: Season totals.
        type: integer
      rolling:
        description: Rolling-window aggregations (e.g. last 10, 20, 30 games).
        items:
          $ref: '#/definitions/core.RunDifferentialWindow'
        type: array
      run_differential:
        description: RS - RA
        type: integer
      runs_allowed:
        type: integer
      runs_scored:
        type: integer
      season:
        type: integer
    type: object
  core.RunDifferentialWindow:
    properties:
      label:
        description: e.g. "last_10", "last_30"
        type: string
      points:
        description: Ordered by window end game/date
        items:
          $ref: '#/definitions/core.RunDifferentialWindowPoint'
        type: array
      window_size:
        description: number of games (e.g. 10, 20, 30)
        type: integer
    type: object
  core.RunDifferentialWindowPoint:
    properties:
      end_date:
        description: YYYY-MM-DD
        type: string
      end_game_id:
        type: string
      games_in_window:
        type: integer
      run_differential:
        description: RS - RA in window
        type: integer
      runs_allowed:
        type: integer
      runs_scored:
        type: integer
    type: object
...
  core.SplitDimension:
    enum:
    - home_away
    - batter_handed
    - pitcher_handed
    - month
    - batting_order
    - inning
    - day_night
    - before_after_asg
    - ground_balls_flys
    type: string
    x-enum-comments:
      SplitDimBatterHanded: vs RHP/LHP by batter side
      SplitDimBattingOrder: lineup spot 1–9
      SplitDimBeforeAfterASG: before/after all-star break
      SplitDimDayNight: day vs night games
      SplitDimGroundBallsFlys: batted ball type
      SplitDimHomeAway: home vs away
      SplitDimInning: inning number
      SplitDimMonth: calendar or season month
      SplitDimPitcherHanded: pitcher side
    x-enum-descriptions:
    - home vs away
    - vs RHP/LHP by batter side
    - pitcher side
    - calendar or season month
    - lineup spot 1–9
    - inning number
    - day vs night games
    - before/after all-star break
    - batted ball type
    x-enum-varnames:
    - SplitDimHomeAway
    - SplitDimBatterHanded
    - SplitDimPitcherHanded
    - SplitDimMonth
    - SplitDimBattingOrder
    - SplitDimInning
    - SplitDimDayNight
    - SplitDimBeforeAfterASG
    - SplitDimGroundBallsFlys
  core.SplitEntityType:
    enum:
    - player
    - team
    type: string
    x-enum-varnames:
    - SplitEntityPlayer
    - SplitEntityTeam
  core.SplitGroup:
    properties:
      ab:
        type: integer
      avg:
        type: number
      bb:
        type: integer
      games:
        type: integer
      h:
        type: integer
      hr:
        type: integer
      key:
        description: e.g. "home", "away", "vs_LHP", "04" (April), "1" (leadoff)
        type: string
      label:
        description: human-readable label
        type: string
      meta:
        additionalProperties:
          type: string
        description: 'arbitrary: month_number, handedness, etc.'
        type: object
      obp:
        type: number
      ops:
        type: number
      pa:
        type: integer
      slg:
        type: number
      so:
        type: integer
    type: object
  core.SplitResult:
    properties:
      dimension:
        allOf:
        - $ref: '#/definitions/core.SplitDimension'
        description: home_away, month, etc.
      entity_id:
        description: PlayerID or TeamID
        type: string
      entity_type:
        allOf:
        - $ref: '#/definitions/core.SplitEntityType'
        description: player or team
      groups:
        description: one per split bucket
        items:
          $ref: '#/definitions/core.SplitGroup'
        type: array
      season:
        type: integer
    type: object
...
  core.Streak:
    properties:
      end_date:
        type: string
      end_game_id:
        type: string
      entity_id:
        description: PlayerID or TeamID (as string)
        type: string
      entity_type:
        allOf:
        - $ref: '#/definitions/core.StreakEntityType'
        description: player or team
      id:
        description: internal identifier for the streak
        type: string
      kind:
        allOf:
        - $ref: '#/definitions/core.StreakKind'
        description: hitting, scoreless_innings
      label:
        description: human-readable label
        type: string
      length:
        description: |-
          Length of the streak:
          - For hitting: number of consecutive games with ≥1 hit.
          - For scoreless innings: total consecutive innings without allowing a run.
        type: integer
      season:
        description: season year (e.g. 2024)
        type: integer
      start_date:
        description: ISO-8601 dates (YYYY-MM-DD) to keep JSON lightweight and frontend-friendly.
        type: string
      start_game_id:
        description: Game/inning bounds of the streak.
        type: string
      timeline:
        description: Timeline for charts or detailed views.
        items:
          $ref: '#/definitions/core.StreakPoint'
        type: array
    type: object
  core.StreakEntityType:
    enum:
    - player
    - team
    type: string
    x-enum-varnames:
    - StreakEntityPlayer
    - StreakEntityTeam
  core.StreakKind:
    enum:
    - hitting
    - scoreless_innings
    type: string
    x-enum-comments:
      StreakKindHitting: 'Hitting streak: consecutive games with at least one hit.'
      StreakKindScorelessInnings: Scoreless innings streak for a pitcher or team staff.
    x-enum-descriptions:
    - 'Hitting streak: consecutive games with at least one hit.'
    - Scoreless innings streak for a pitcher or team staff.
    x-enum-varnames:
    - StreakKindHitting
    - StreakKindScorelessInnings
  core.StreakPoint:
    properties:
      ab:
        type: integer
      date:
        description: YYYY-MM-DD
        type: string
      game_id:
        type: string
      h:
        type: integer
      index:
        description: Index within the streak (1..Length).
        type: integer
      ip:
        description: For scoreless innings streaks.
        type: number
      pa:
        description: For hitting streaks.
        type: integer
      runs_allowed:
        type: integer
    type: object
...
  core.WinExpectancy:
    properties:
      created_at:
        type: string
      end_year:
        type: integer
      id:
        type: integer
      inning:
        type: integer
      is_bottom:
        type: boolean
      outs:
        type: integer
      runners_state:
        type: string
      sample_size:
        description: Number of historical games used
        type: integer
      score_diff:
        type: integer
      start_year:
        type: integer
      updated_at:
        type: string
      win_probability:
        description: 0.0 to 1.0
        type: number
    type: object
  core.WinExpectancyEra:
    properties:
      end_year:
        type: integer
      label:
        description: e.g., "Modern Era", "Steroid Era"
        type: string
      start_year:
        type: integer
      state_count:
        description: Number of unique game states in this era
        type: integer
      total_sample:
        description: Total sample size across all states
        type: integer
    type: object
  core.WinProbabilityCurve:
    properties:
      away_team:
        type: string
      game_id:
        type: string
      home_team:
        type: string
      points:
        description: One point per play/event, ordered by EventIndex ascending.
        items:
          $ref: '#/definitions/core.WinProbabilityPoint'
        type: array
      season:
        type: integer
    type: object
  core.WinProbabilityPoint:
    properties:
      away_score:
        type: integer
      away_win_prob:
        type: number
      bases:
        description: |-
          Simple base-state encoding
          Example: "100" = runner on 1st only, "011" = runners on 2nd & 3rd.
        type: string
      description:
        description: e.g. "Trout homers to LF (fly ball)."
        type: string
      event_index:
        description: 1..N within the game
        type: integer
      home_score:
        description: Score state after the event.
        type: integer
      home_win_prob:
        description: Win probabilities for each team AFTER the event (0.0–1.0).
        type: number
      inning:
        description: 1..9, 10+
        type: integer
      outs:
        description: 0..3
        type: integer
      top_of_inning:
        type: boolean
    type: object
info:
  contact:
    email: info@stormlightlabs.org
    name: API Support
    url: https://github.com/stormlightlabs/baseball
  description: |
    <!-- markdownlint-disable -->

    A comprehensive REST API for baseball statistics serving data from the Lahman Baseball Database and Retrosheet

    ## Attributions

    **[SABR Lahman Database](https://sabr.org/lahman-database/)**: The information used here was obtained free of charge from and is copyrighted by Sean Lahman.

    **[Retrosheet.org](https://www.retrosheet.org/)**: The information used here was obtained free of charge from and is copyrighted by Retrosheet.

    wOBA weights and seasonal constants from **[FanGraphs Guts!](https://www.fangraphs.com/tools/guts)**, used under FanGraphs’ published terms.
  license:
    name: MPL-2.0
    url: https://opensource.org/license/mpl-2-0
  title: Baseball API
  version: "1.0"
paths:
...
  /ejections:
    get:
      consumes:
      - application/json
      description: Get ejections with optional filters for player, umpire, team, and
        role
      parameters:
      - description: Filter by ejected player ID
        in: query
        name: player_id
        type: string
      - description: Filter by umpire ID
        in: query
        name: umpire_id
        type: string
      - description: Filter by team
        in: query
        name: team
        type: string
      - description: Filter by role
        enum:
        - P
        - M
        - C
        in: query
        name: role
        type: string
      - description: Filter by season year
        in: query
        name: season
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List ejections
      tags:
      - ejections
...
  /games/{game_id}/win-probability:
    get:
      consumes:
      - application/json
      description: Get play-by-play win probability for a game
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.WinProbabilityCurve'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get game win probability curve
      tags:
      - derived
      - games
...
  /players/{player_id}/splits:
    get:
      consumes:
      - application/json
      description: Get batting statistics split by dimension (home/away, vs handedness,
        month, etc.)
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - description: 'Split dimension: home_away, pitcher_handed, or month'
        in: query
        name: dimension
        required: true
        type: string
      - description: Season year
        in: query
        name: season
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.SplitResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get player batting splits
      tags:
      - derived
      - players
...
  /players/{player_id}/streaks:
    get:
      consumes:
      - application/json
      description: Get hitting or scoreless innings streaks for a player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: string
      - description: 'Streak kind: hitting or scoreless_innings'
        in: query
        name: kind
        required: true
        type: string
      - description: Season year
        in: query
        name: season
        required: true
        type: integer
      - default: 5
        description: Minimum streak length
        in: query
        name: min_length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Streak'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get player streaks
      tags:
      - derived
      - players
...
  /seasons/{year}/ejections:
    get:
      consumes:
      - application/json
      description: Get all ejections that occurred during a specific season
      parameters:
      - description: Season year
        in: path
        name: year
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get ejections for a specific season
      tags:
      - ejections
      - seasons
...
  /teams/{team_id}/run-differential:
    get:
      consumes:
      - application/json
      description: Get season run differential with rolling windows for a team
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - description: Season year
        in: query
        name: season
        required: true
        type: integer
      - default: '"10,20,30"'
        description: Comma-separated rolling window sizes (e.g., 10,20,30)
        in: query
        name: windows
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.RunDifferentialSeries'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get team run differential
      tags:
      - derived
      - teams
...
  /win-expectancy:
    get:
      consumes:
      - application/json
      description: Get the historical win probability for a specific game situation
      parameters:
      - description: Inning (1-9)
        in: query
        name: inning
        required: true
        type: integer
      - description: Bottom of inning (true/false)
        in: query
        name: is_bottom
        required: true
        type: boolean
      - description: Outs (0-2)
        in: query
        name: outs
        required: true
        type: integer
      - description: Runners state (e.g., ___, 1__, 12_, 123)
        in: query
        name: runners
        required: true
        type: string
      - description: Score differential from home team perspective (-11 to +11)
        in: query
        name: score_diff
        required: true
        type: integer
      - description: Start year for historical era filter
        in: query
        name: start_year
        type: integer
      - description: End year for historical era filter
        in: query
        name: end_year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.WinExpectancy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get win expectancy for a game state
      tags:
      - derived
      - win-expectancy
  /win-expectancy/eras:
    get:
      consumes:
      - application/json
      description: Get all available historical eras in the win expectancy database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.WinExpectancyEra'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List available win expectancy eras
      tags:
      - derived
      - win-expectancy
swagger: "2.0"
...